python-flask-app/requirements.txt
Flask==2.0.3
Flask-CORS==3.0.10
pymongo==4.3.3
Werkzeug==2.0.3
bcrypt==3.2.0
PyJWT==2.3.0